requirements.txt
discordpy
dotenv